Requested Changes + UI Fixes
