+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXHeader/PXHeaderView + Tracking.swift
@@ -1,0 +1,12 @@
+//
+//  PXHeaderView + Tracking.swift
+//  MercadoPagoSDK
+//
+//  Created by AUGUSTO COLLERONE ALFONSO on 21/02/2019.
+//
+
+import Foundation
+
+protocol PXHeaderTrackingProtocol: NSObjectProtocol {
+    func didTapCloseButton()
+}
